backend/requirements.txt
# Web Framework
fastapi==0.112.2
uvicorn[standard]==0.30.6

# Database
sqlalchemy==2.0.35
alembic==1.13.2
pymysql==1.1.1

# HTTP & Web Scraping
httpx==0.27.2
beautifulsoup4==4.12.3
lxml==5.3.0

# Configuration & Validation
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1

# Data Processing
pandas==2.2.3

# Utilities
tenacity==9.0.0
loguru==0.7.2
cryptography==43.0.1

# AI & External Services
openai==1.107.1
pyairtable==2.3.3

# Google Cloud (BigQuery)
google-cloud-bigquery==3.25.0
db-dtypes==1.3.0



